Override the #find_free_port method in ruby-debug-ide
As similarly proposed in https://github.com/ruby-debug/ruby-debug-ide/issues/107
This hack is necessary as long as https://github.com/ruby-debug/ruby-debug-ide/pull/164 is not merged and released because:
* We cannot easily modify the Gemfile because it is not in the same repository. So we need to use `gem install` in the Dockerfile to do it
* The newest RubyMine requires a prerelease version of ruby-debug-ide
* The fix with SO_REUSEPORT is only available on a specific branch of a specific fork, which is not up-to-date with the prerelease version
* We cannot use the monkey-patch in an initializer as proposed in the issue, because wagon:test doesn't run Rails and thus doesn't consider the initializers

The combination of these three aspects renders it impossible to easily install a version that includes the fix.
So instead, we simply overwrite the relevant code inside the bundle with something that works for our case for now.
As soon as the PR is merged and released, we can revert to installing the official, unaltered version into our docker container.